TextureAsset: changing image size to uint16 (256 cannot fit in uint8, duh)

+++ include/base/tundra/assets/texture/texture-asset.hpp
@@ -18,8 +18,8 @@ namespace td {
 
         TextureMode mode = TextureMode::TrueColor;
 
-        uint8 pixels_width = 0;
-        uint8 pixels_height = 0;
+        uint16 pixels_width = 0;
+        uint16 pixels_height = 0;
 
         TextureAssetLoadInfo* load_info = nullptr;
